requirements.txt
pyautogui==0.9.54
pyobjc-framework-Quartz==12.1
opencv-python==4.10.0.84
numpy==2.1.3
Pillow==11.0.0
pytest==8.3.3
